LPS-56211 Move code from blogs actions into upload handlers and delegate
